scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile ./scripts/requirements.in
#
asyncio==3.4.3
    # via -r ./scripts/requirements.in
azure-common==1.1.28
    # via
    #   azure-mgmt-authorization
    #   azure-mgmt-resource
azure-core==1.29.7
    # via
    #   azure-identity
    #   azure-mgmt-core
azure-identity==1.15.0
    # via -r ./scripts/requirements.in
azure-mgmt-authorization==4.0.0
    # via -r ./scripts/requirements.in
azure-mgmt-core==1.4.0
    # via
    #   azure-mgmt-authorization
    #   azure-mgmt-resource
azure-mgmt-resource==23.0.1
    # via -r ./scripts/requirements.in
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.1
    # via
    #   azure-identity
    #   msal
    #   pyjwt
idna==3.6
    # via requests
isodate==0.6.1
    # via
    #   azure-mgmt-authorization
    #   azure-mgmt-resource
msal==1.26.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
packaging==23.2
    # via msal-extensions
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
python-dotenv==1.0.1
    # via -r ./scripts/requirements.in
requests==2.31.0
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   isodate
typing-extensions==4.9.0
    # via azure-core
urllib3==2.1.0
    # via requests
